Sync all events, not just upcoming

+++ index.js
@@ -2,55 +2,51 @@ const cal = require('./cal');
 const squarespace = require('./squarespace');
 const log = require('./log');
 
-function areEventsEqual(squarespaceEvent, googleCalendarEvent) {
-  return squarespaceEvent.title === googleCalendarEvent.summary &&
-    cal.formatDateTime(squarespaceEvent.startDate, squarespaceEvent.startTime) === googleCalendarEvent.start.dateTime &&
-    cal.formatDateTime(squarespaceEvent.endDate, squarespaceEvent.endTime) === googleCalendarEvent.end.dateTime &&
-    squarespaceEvent.address === googleCalendarEvent.locator &&
-    squarespaceEvent.description === googleCalendarEvent.description &&
-    squarespaceEvent.href === googleCalendarEvent.source.url;
-}
-
-async function syncEventsInGoogleCalender(latestSquarespaceEvents) {
+async function syncEventsInGoogleCalender(squarespaceEvents) {
   const changes = {
     toDelete: [],
     toCreate: [],
   };
 
   const existingGoogleEvents = await cal.getAllEvents();
+  log.debugJSON(existingGoogleEvents);
 
   // Find new squarespace events to add
-  for (const latestSquarespaceEvent of latestSquarespaceEvents) {
-    let found = false;
-    for (const existingGoogleEvent of existingGoogleEvents.items) {
-      if (areEventsEqual(latestSquarespaceEvent, existingGoogleEvent)) {
-        found = true;
-        break;
-      }
-    }
-    if (!found) {
-      changes.toCreate.push(latestSquarespaceEvent);
-    }
+  for (const squarespaceEvent of squarespaceEvents) {
+    // let found = false;
+    // for (const existingGoogleEvent of existingGoogleEvents.items) {
+    //   if (cal.areEventsEqual(squarespaceEvent, existingGoogleEvent)) {
+    //     found = true;
+    //     break;
+    //   }
+    // }
+    // if (!found) {
+      changes.toCreate.push(squarespaceEvent);
+    // }
   }
 
   // Find existing Google Calendar events to remove
-  for (const existingGoogleEvent of existingGoogleEvents.items) {
-    let found = false;
-    for (const latestSquarespaceEvent of latestSquarespaceEvents) {
-      if (areEventsEqual(latestSquarespaceEvent, existingGoogleEvent)) {
-        found = true;
-        break;
-      }
-    }
-    if (!found && new Date(existingGoogleEvent.start.dateTime) > new Date()) {
+  for (const existingGoogleEvent of existingGoogleEvents) {
+    // let found = false;
+    // for (const squarespaceEvent of squarespaceEvents) {
+    //   if (cal.areEventsEqual(squarespaceEvent, existingGoogleEvent)) {
+    //     found = true;
+    //     break;
+    //   }
+    // }
+    // if (!found) {
       changes.toDelete.push(existingGoogleEvent);
-    }
+    // }
   }
 
   log.debugJSON(changes);
 
   for (const e of changes.toDelete) {
-    log.info('Google Calendar delete event:', e.summary);
+    log.info('Google Calendar delete event:', e.summary, e.creator.email);
+
+    if (e.creator.email !== cal.creator) {
+      throw new Error(`Will only delete events created by ${cal.creator}`);
+    }
     const result = await cal.deleteEvent(e.id);
     log.debugJSON(result);
   }
@@ -63,7 +59,7 @@ async function syncEventsInGoogleCalender(latestSquarespaceEvents) {
 }
 
 async function main() {
-  const events = await squarespace.getEvents();
+  const events = await squarespace.getAllEvents();
   await syncEventsInGoogleCalender(events);
 }
